requirements.txt
# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
playwright==1.40.0

# AI Integration
openai==1.3.5
langchain==0.0.340
transformers==4.35.2
torch==2.1.1

# Test Framework
pytest==7.4.3
pytest-html==4.1.1
pytest-bdd==7.0.0
behave==1.2.6

# Data Processing
pandas==2.1.3
numpy==1.24.3
lxml==4.9.3

# Reporting
jinja2==3.1.2
plotly==5.17.0
dash==2.14.2

# Configuration
pydantic==2.5.0
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging
loguru==0.7.2

# HTTP Client for AI APIs
httpx==0.25.2
aiohttp==3.9.1

# JSON Schema
jsonschema==4.19.2

# Image processing (for screenshots)
Pillow==10.1.0 